backend/requirements.txt
Flask
Flask-Cors
faiss-cpu
PyPDF2
numpy
sentence-transformers
gunicorn
